Move generateRuntimeFilters.ts to helpers dir
